Merge new logging API based on boost::log
This commit merges a new logging API based on boost::log which works similar
to the old LoggingBase template.

The build for this can be disabled by setting dsnutil_cpp_WITH_BOOST_LOG to OFF
during Makefile generation. If the option is set to ON it'll add Boost >= 1.54
to the build dependencies and build a new library "dsnutil_cpp-log" that
contains basic initialization code for boost::log.

Signed-off-by: Peter Hille (png!das-system) <peter@das-system-networks.de>